testui/requirements.txt
click==8.3.0
colorama==0.4.6
dacite==1.9.2
packaging==25.0
pyqt-toast-notification==1.3.3
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.10.2
python-dotenv==1.1.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
QtPy==2.4.3
